+++ myokit/tests/test_config.py
@@ -38,6 +38,26 @@ lib = five;six
 inc = three;eight
 """
 
+# Config with empty paths and spaces
+config_empties_1 = """
+[sundials]
+lib = five
+inc = three;four;
+[opencl]
+lib = one;;   two point five;three;
+inc =
+"""
+
+# Config with empty paths and " ;", which in Python 2 is ignored
+config_empties_2 = """
+[sundials]
+lib = five ; six
+inc = three;four;
+[opencl]
+lib = one ;;   two point five ;three;
+inc =
+"""
+
 # Qt options
 config_pyside = """
 [gui]
@@ -144,6 +164,41 @@ class TestConfig(unittest.TestCase):
                 self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
                 self.assertEqual(myokit.OPENCL_LIB, ['five', 'six'])
                 self.assertEqual(myokit.OPENCL_INC, ['three', 'eight'])
+
+                # Lists of paths should be filtered for empty values and
+                # trimmed
+                myokit.SUNDIALS_LIB = []
+                myokit.SUNDIALS_INC = []
+                myokit.OPENCL_LIB = []
+                myokit.OPENCL_INC = []
+                with open(d.path('myokit.ini'), 'w') as f:
+                    f.write(config_empties_1)
+                config._load()
+                self.assertEqual(myokit.SUNDIALS_LIB, ['five'])
+                self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
+                self.assertEqual(
+                    myokit.OPENCL_LIB,
+                    ['one', 'two point five', 'three'])
+                self.assertEqual(myokit.OPENCL_INC, [])
+
+                # Even if the list contains " ;", which Python 2's config
+                # parser treats as a comment
+                myokit.SUNDIALS_LIB = []
+                myokit.SUNDIALS_INC = []
+                myokit.OPENCL_LIB = []
+                myokit.OPENCL_INC = []
+                with open(d.path('myokit.ini'), 'w') as f:
+                    f.write(config_empties_2)
+                if sys.hexversion < 0x03020000:
+                    self.assertRaises(ImportError, config._load)
+                else:
+                    config._load()
+                    self.assertEqual(myokit.SUNDIALS_LIB, ['five', 'six'])
+                    self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
+                    self.assertEqual(
+                        myokit.OPENCL_LIB,
+                        ['one', 'two point five', 'three'])
+                    self.assertEqual(myokit.OPENCL_INC, [])
 
                 # Qt gui options
                 with open(d.path('myokit.ini'), 'w') as f: